Added the drag and drop feature module
